+++ Sources/PointFree/Blog/BlogPostIndex.swift
@@ -9,6 +9,7 @@ import Prelude
 import Styleguide
 import Tuple
 import View
+import Views
 
 let blogIndexMiddleware: (Conn<StatusLineOpen, Tuple3<User?, SubscriberState, Route?>>) -> IO<Conn<ResponseEnded, Data>> =
   writeStatus(.ok)
