+++ Sources/PointFree/Blog/BlogPostShow.swift
@@ -11,6 +11,7 @@ import Prelude
 import Styleguide
 import Tuple
 import View
+import Views
 
 let blogPostShowMiddleware: Middleware<StatusLineOpen, ResponseEnded, Tuple4<Either<String, BlogPost.Id>, User?, SubscriberState, Route?>, Data> =
   filterMap(
