+++ Sources/PointFree/Emails/NewBlogPostEmail.swift
@@ -11,6 +11,7 @@ import PointFreeRouter
 import Prelude
 import Styleguide
 import View
+import Views
 
 let newBlogPostEmail = simpleEmailLayout(newBlogPostEmailContent)
   .contramap { post, subscriberAnnouncement, nonSubscriberAnnouncement, user in
