+++ Sources/PointFree/Home.swift
@@ -12,6 +12,7 @@ import Styleguide
 import Tuple
 import UrlFormEncoding
 import View
+import Views
 
 let homeMiddleware: Middleware<StatusLineOpen, ResponseEnded, Tuple3<User?, SubscriberState, Route?>, Data> =
   writeStatus(.ok)
