+++ Sources/PointFree/SiteMiddleware.swift
@@ -216,65 +216,3 @@ private let allowedInsecureHosts: [String] = [
   "localhost"
 ]
 
-enum SubscriberState {
-  case nonSubscriber
-  case owner(hasSeat: Bool, status: Stripe.Subscription.Status)
-  case teammate(status: Stripe.Subscription.Status)
-
-  init(user: User?, subscription: Models.Subscription?) {
-    switch (user, subscription) {
-    case let (.some(user), .some(subscription)):
-      if subscription.userId == user.id {
-        self = .owner(
-          hasSeat: user.subscriptionId != nil,
-          status: subscription.stripeSubscriptionStatus
-        )
-      } else {
-        self = .teammate(status: subscription.stripeSubscriptionStatus)
-      }
-
-    case (.none, _), (.some, _):
-      self = .nonSubscriber
-    }
-  }
-
-  var status: Stripe.Subscription.Status? {
-    switch self {
-    case .nonSubscriber:
-      return nil
-    case let .owner(_, status):
-      return status
-    case let .teammate(status):
-      return status
-    }
-  }
-
-  var isActive: Bool {
-    return self.status == .some(.active)
-  }
-
-  var isPastDue: Bool {
-    return self.status == .some(.pastDue)
-  }
-
-  var isOwner: Bool {
-    if case .owner = self { return true }
-    return false
-  }
-
-  var isTeammate: Bool {
-    if case .teammate = self { return true }
-    return false
-  }
-
-  var isNonSubscriber: Bool {
-    if case .nonSubscriber = self { return true }
-    return false
-  }
-
-  var isActiveSubscriber: Bool {
-    if case .teammate(status: .active) = self { return true }
-    if case .owner(hasSeat: true, status: .active) = self { return true }
-    return false
-  }
-}
