+++ Sources/PointFree/Tasks.swift
@@ -7,6 +7,7 @@ import PointFreePrelude
 import Prelude
 import Styleguide
 import View
+import Views
 
 public func sendWelcomeEmails() -> EitherIO<Error, Prelude.Unit> {
   let zippedEmails = zip3(
